splitup benchmarks. switch to fluent config
