+++ src/ef-json-query-testing/Benchmarks/BoolBenchmarks.cs
@@ -1,0 +1,40 @@
+﻿using BenchmarkDotNet.Attributes;
+using Bogus;
+using ef_json_query_testing.Enums;
+using ef_json_query_testing.Models;
+
+namespace ef_json_query_testing.Benchmarks
+{
+    internal class BoolBenchmarks : BaseBenchmark
+    {
+        public IEnumerable<object[]> BenchmarkData_BoolMatch()
+        {
+            var boolTypeField = Context.DynamicFields.First(m => m.DataType == DataTypes.BoolValue);
+
+            var faker = new Faker();
+            var item = faker.PickRandom<DynamicMediaInformation>(Context.DynamicMediaInformation.Where(m => m.FieldId == boolTypeField.DynamicFieldId));
+
+            yield return new object[] { item.FieldId, item.Value };
+        }
+
+        [Benchmark]
+        [BenchmarkCategory("json", "bool", "raw")]
+        [ArgumentsSource(nameof(BenchmarkData_BoolMatch))]
+        public void Benchmark_Bool_JSON_Raw(int i, string str) => Search.JsonSearch_Raw(i, str);
+
+        [Benchmark]
+        [BenchmarkCategory("json", "bool", "magic")]
+        [ArgumentsSource(nameof(BenchmarkData_BoolMatch))]
+        public void Benchmark_Bool_JSON_Magic(int i, string str) => Search.JsonSearch_EfMagic(i, str);
+
+        [Benchmark]
+        [BenchmarkCategory("table", "bool", "info")]
+        [ArgumentsSource(nameof(BenchmarkData_BoolMatch))]
+        public void Benchmark_Bool_Table_Info(int i, string str) => Search.TableSearch_Info(i, str);
+
+        [Benchmark]
+        [BenchmarkCategory("table", "bool", "media")]
+        [ArgumentsSource(nameof(BenchmarkData_BoolMatch))]
+        public void Benchmark_Bool_Table_Media(int i, string str) => Search.TableSearch_Media(i, str);
+    }
+}
